Add non-Go-user friendly build instructions
Go is not  super helpful if you checkout the code to a random path, then try and follow the build instructions in the README. Add some instructions that help users who don't deal with Go on a day-to-day basis to get it right.

Fixes #1070. Thanks to @johnbelamaric for reminding me about Go's deficiencies.