Use `amd64` instread of `x86_64` when building architecture
In Makefile.release `shell uname -m` was used to find the current architecture.
On amd64 systems this might be reported as `x86_64` depending on the Linux.
However, in order to have manifest docker images we need to stick with
golang's architecture tag of `amd64`.

This fix changes instead so that the correct architecture conforming to
golang (`amd64`) could be build.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>