Fix version for apimachinery and ugorji/go in Gopkg.toml
This fix is an attempt to fix the issue raised #1192.
The issue that a plain `dep -ensure update` fails is because
the incompatibility of ugorji/go with the rest of packages (etcd, k8s etc)
and the incompatibility of client-go vs. apimachinery.

This fix fixed apimachinery and ugorji/go so that it is possible
to use `dep -ensure update`.

Previously updates has to be done individually, e.g.,
`dep -ensure update github.com/coreos/etcd`

This fix fixes #1192.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>