CoreDNS distributed plugin with NSID
This fix tries to implement a CoreDNS distributed in a
cluster, and each CoreDNS node could return the
domain names specified by its peers (through NSID).

The basic idea is that each node will use NSID to publish
its domain name. The cluster is specified through the list
of IPs or the list of subnets. When every node is up,
it will send NSID queries to every IP in the list, and receive
the domain names by response. If the domain name is within
the specified domain, then each node will expose to the
outside. This is more of an attempt to assign
each node within a cluster a unique name in a distributed fashion,
without relying on external storages or K-V stores.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>