plugin/cache: remove goroutine from prefetch handler
We used to start the prefetch in a goroutine, but because the responsewriter is closed
by coredns when this handler returns, this leads to a crash when any other plugin
accesses the (now partly closed) response writer.

Optionally we can create a new ResponseWriter that fakes a few methods
to make it work, but this quickly becomes a larger refacter.

The downside is that the unfortunate query that triggers the prefetch
will see inflated duration.