Refactor parseForward to ensure all servers have a valid healthcheck
The previous logic relied on the length of the TO argument matching the
created proxies in order to determine which proxies did and did not use
TLS. This fell apart in cases where TO was a reference to a resolv.conf
file containing more than one DNS server.

Using 'forward . /etc/resolv.conf' and a resolv.conf with two nameserver
entries would result in the first nameserver using protocol type DNS and
the second using protocol type nil, resulting in a HealthCheck that would
panic when called. I assume using the same resolv.conf with something like
'forward . /etc/resolv.conf tls://127.0.0.1' would result in the second
nameserver in resolv.conf being incorrectly configured to use TLS and the
explicitly configured 127.0.0.1 nameserver *not* using TLS (and in fact
having a nil protocol health check).