plugin/etcd: Fix CNAME query for empty host field
When a CNAME is requested, and a service with a filled out host field is
not present for the exact key, a NODATA result should be returned. If
the key exists, and the host field is empty, coredns erroneously returns
a NOERROR response with a "." as the result. This would for example
manifest when the requested key had a service representing a TXT record
(Text field filled, empty Host field).

Fixing by specifically checking for empty Host fields when a CNAME
lookup is done, and returning an empty answer, with the NOERROR Rcode.