Remove version pinning of thrift, ugoriji/go, and etcd
For incompatibility reasons at one point, we were forced
to pining the version of thrift, ugoriji/go, and etcd
to very specific versions (some are not versioned commits)
to get around the build issues.

It looks like those incompatibility issues are gone so
we could remove the pinning of thrift, ugoriji/go, and etcd.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>