Metric for measuring DNS Programming Latency SLI.
The DNS Programming Latency definition can be found [here](https://github.com/kubernetes/community/blob/master/sig-scalability/slos/dns_programming_latency.md)

This PR covers only "headless with selector" services, other service
kinds are blocked on the impossibility of determining the LastUpdateTime of a
service object.

The PR bears some similarities to
https://github.com/kubernetes/kubernetes/pull/71999, which introdced
In-Cluster Network Programming Latency. The main difference is that
there is no actual "programming" happening in CoreDNS (for comparison,
in kube-proxy the network programming consists of writing IPTables/IPVS
rules). The CoreDNS serves the content directly from the endpoints/service/pod cache,
creating DNS records on the fly. Thus, we assume that the programming of DNS ends in the
moment when the endpoints/service/pod change reaches the CoreDNS via the
Watch mechanism.