plugin/forward: bring back jitter
Very minimal jitter implementation that adds up to 25% jitter to the
expiration for the connection. This helps when we see a burst in
connections which all race to get a TLS conn and then also are expired
at the same time, meaning we can see a repeat.

Another thing could be to keep a couple of connection in the cache to be
able to serve right away (not implemented in this PR)

Signed-off-by: Miek Gieben <miek@miek.nl>