plugin/hosts: send nxdomain response
As specified in the plugin documentation, CoreDNS assumes a plugin has sent a
response in all cases except SERVFAIL, REFUSED, FORMERR, and NOTIMP. The
hosts plugin was returning NXDOMAIN but sending no response in the case where
it didn't have any records for the qname, and fallthrough was not available.
This resulted in no response being sent to the client.