[test/responsewriter] provide mechanism to specify custom Remote IP address and number of message written
This is useful to build test when the behaviour may be expected based on client IP (Geo IP and the likes)

GetWriteMsgCallCount() make it easier to validate if WriteMsg was called or not.

request/request_test.go contains a test validating the custom IP.
plugin/whoami/whoami_test.go test was modified to validate the
GetWriteMsgCallCount() function.