startup: add logo
As discussed in #3225, lets add a little logo. This PR incorperates this
in the startup text. It also removes the logging output as this outputs
identical lines of text, but for no clear reason. This means --quiet now
means "no output on startup at all".

Currently it looks like this:

~~~
.:1053
   ______                ____  _   _______
  / ____/___  ________  / __ \/ | / / ___/	CoreDNS-1.6.2
 / /   / __ \/ ___/ _ \/ / / /  |/ /\__ \ 	linux/amd64, go1.12.9,
/ /___/ /_/ / /  /  __/ /_/ / /|  /___/ /
\____/\____/_/   \___/_____/_/ |_//____/
~~~

We have 2 lines extra on the right if we need to print more (no ideas
come to mind currently).

Signed-off-by: Miek Gieben <miek@miek.nl>