plugin/forward: remove the channel communication
Use three, protected by a mutex, slices of persistent connections. Just
pick one from the list. This removes the send and then receive; the
start/stop dance and simplifies the code.

This still needs to be load tested.

Possible things todo and fix is set a max on the number of conns in the
slice; 1000 each?

Another thing is the following scenario: big spike in qps, slice full,
then nothing for 5s; then all conns in the slice are too old to be used
and we will redial for the first 1000; is this problem? Or is this only
a problem when low qps (then it's not a problem).

Signed-off-by: Miek Gieben <miek@miek.nl>