presubmit: test README.md sections
Again one of those things that's super manual. Enforce the standard we
have and fix the few READMEs that didn't adhere to it.

For *transfer* I left a TODO in the (new) Examples section.

plugin.md has been saying this should be done for ages - but without
checks things slip through - as apparent in this PR.

Signed-off-by: Miek Gieben <miek@miek.nl>