Remove grobie from maintainers
I haven't been maintaining production DNS setups in quite a while and
don't have the time to review CoreDNS pull requests. Hereby I resign
from all positions in the CoreDNS project.

CoreDNS is a great project and SoundCloud still processes hundreds of
thousands of requests per second with the help of CoreDNS. Thanks a lot
for your hard work and all the best for the future.

Signed-off-by: Tobias Schmidt <tobidt@gmail.com>