plugin/trace: Fix zipkin json_v2
In #4171 i noticed that the client_server was not being used anymore.
It appears that it should still be there.

While testing that change i noticed that the zipkin provider was borked and couldn't push traces anymore
Tcpdump showed that the zipkin server was responding with

```
Expected a JSON_V1 encoded list, but received: JSON_V2
```

There was also an issue (#4160)  with
```

	recorder, err := zipkin.NewEndpoint(t.serviceName, t.serviceEndpoint)
```
As `t.serviceEndpoint` is `:53` in most cases
Zipkin can't parse the ip:port tuple and send a warning with
```
[WARNING] build Zipkin endpoint found err: lookup : no such host
```
cf: [openzipkin/endpoint.go](https://github.com/openzipkin/zipkin-go/blob/master/endpoint.go#L27-L80)

To actively set a correct values in serviceEnpoint you need to use the bind plugin
```
.:53 {
    bind 127.0.0.1
    trace
}
```

Signed-off-by: bob <bob.bouteillier@datadoghq.com>