plugin/dnstap: remove buffered channel
This started out as a way to remove the buffered channel, but as I
progressed I noticed that dnstap_encoder.go basically redos a bunch
of dnstap stuff; it's essentially not needed; so I removed it. This
revealed other issues as concurrency on the internal buffer maintained,
so that was also removed.

The bidirectional stuff didn't work, but this was masked by the fact
everything was buffered - removed that too.

Might a whole bunch of methods private as they are not needed outside
this package.

Signed-off-by: Miek Gieben <miek@miek.nl>