[plugin][forward] - Populate return code on upstream error.
Following an incident we noticed that the global metrics did not have the right return code when upstream failure were happening.
More specifically `i/o timeout` errors
Also if the client is patient enough (>30s) the query is returned with a *noerror* code.

![Graph showing app errors](https://user-images.githubusercontent.com/5513509/104312937-989c1600-54d7-11eb-9295-736fb9e5d900.png)
![Graph showing apm errors](https://user-images.githubusercontent.com/5513509/104312798-668ab400-54d7-11eb-863d-33232b565e4b.png)

![Graph showing no servfail errors](https://user-images.githubusercontent.com/5513509/104313210-fd577080-54d7-11eb-8044-1745de45ba1d.png)
☝️ Blue is NOERROR , which should have been another color because of a different error code.

In order to fix that particular issue, i though of 2 ways to go about it.
The one in the PR
Or modifying the [metrics/handler.go:ServeDns](https://github.com/coredns/coredns/blob/master/plugin/metrics/handler.go#L29) function to use the `status` instead of `rw.Rcode`

Signed-off-by: bob <bob.bouteillier@datadoghq.com>