workflow: guard against faillint failing
So there is (and will always) be a disconnect between latest Go in the
workflow and the actual latest Go used by other utils. Faillint move to
go 1.16 features and for some reason this now borkes.

Add ||true to allow failling to not compile, this will skip that
particular test (so be it)

Signed-off-by: Miek Gieben <miek@miek.nl>